# FrontEnd/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification - Passluk",
    "description": "OpenApi documentation for Spring Security",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Passluk",
      "email": "BreadExchange@gmail.com"
    },
    "license": {
      "name": "Licence BradExchange",
      "url": "https://BreadExchange.com"
    },
    "version": "2.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/user/upload-profile-picture": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "operationId": "uploadProfilePicture",
        "parameters": [
          {
            "name": "direction",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/addAddress": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "operationId": "addAddress",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewAddress"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/bakery/upload-profile-picture": {
      "post": {
        "tags": [
          "bakery-controller"
        ],
        "operationId": "uploadProfilePicture_1",
        "parameters": [
          {
            "name": "idBac",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/bakery/upload-back-picture": {
      "post": {
        "tags": [
          "bakery-controller"
        ],
        "operationId": "uploadBackPicture",
        "parameters": [
          {
            "name": "idBac",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/bakery/updateWeek": {
      "post": {
        "tags": [
          "bakery-controller"
        ],
        "operationId": "updateWeek",
        "parameters": [
          {
            "name": "IdBakery",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyWeekRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/bakery/newItem": {
      "post": {
        "tags": [
          "bakery-controller"
        ],
        "operationId": "newItem",
        "parameters": [
          {
            "name": "idBac",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "idCat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "request"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "request": {
                    "$ref": "#/components/schemas/ItemRequest"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/bakery/newCategory": {
      "post": {
        "tags": [
          "bakery-controller"
        ],
        "operationId": "newCategory",
        "parameters": [
          {
            "name": "bakeryId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "nomeCat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/bakery/modItem": {
      "post": {
        "tags": [
          "bakery-controller"
        ],
        "operationId": "updateItem",
        "parameters": [
          {
            "name": "idItem",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "idBac",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/bakery/addAddress": {
      "post": {
        "tags": [
          "bakery-controller"
        ],
        "operationId": "addAddress_1",
        "parameters": [
          {
            "name": "idBac",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewAddress"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/bakery/updateSocial": {
      "patch": {
        "tags": [
          "bakery-controller"
        ],
        "operationId": "updateSocial",
        "parameters": [
          {
            "name": "idBac",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSocialBakery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/bakery/updateDesc": {
      "patch": {
        "tags": [
          "bakery-controller"
        ],
        "operationId": "updateDesc",
        "parameters": [
          {
            "name": "idBac",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReqModDesc"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/bakery/rejectOrder": {
      "patch": {
        "tags": [
          "bakery-controller"
        ],
        "operationId": "rejectOrder",
        "parameters": [
          {
            "name": "idOrder",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/bakery/changeStatus": {
      "patch": {
        "tags": [
          "bakery-controller"
        ],
        "operationId": "changeStatus",
        "parameters": [
          {
            "name": "idBac",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/bakery/changeStatusOrder": {
      "patch": {
        "tags": [
          "bakery-controller"
        ],
        "operationId": "changeStatusOrder",
        "parameters": [
          {
            "name": "idOrder",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/bakery/acceptOrder": {
      "patch": {
        "tags": [
          "bakery-controller"
        ],
        "operationId": "acceptOrder",
        "parameters": [
          {
            "name": "idOrder",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/bakery/delItem": {
      "delete": {
        "tags": [
          "bakery-controller"
        ],
        "operationId": "delItem",
        "parameters": [
          {
            "name": "idItem",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "idCat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "idBac",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/bakery/delCategory": {
      "delete": {
        "tags": [
          "bakery-controller"
        ],
        "operationId": "delCategory",
        "parameters": [
          {
            "name": "idCat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LocalTime": {
        "type": "object",
        "properties": {
          "hour": {
            "type": "integer",
            "format": "int32"
          },
          "minute": {
            "type": "integer",
            "format": "int32"
          },
          "second": {
            "type": "integer",
            "format": "int32"
          },
          "nano": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "WeekDay": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "dayOfWeek": {
            "type": "string"
          },
          "morningOpening": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "morningClosing": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "eveningOpening": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "eveningClosing": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "closed": {
            "type": "boolean"
          },
          "allDayOpen": {
            "type": "boolean"
          }
        }
      },
      "NewAddress": {
        "required": [
          "city",
          "country",
          "name",
          "number",
          "postalCode",
          "provincia",
          "state",
          "street",
          "telNumber"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "provincia": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "number": {
            "type": "string"
          },
          "telNumber": {
            "type": "string"
          }
        }
      },
...
      "ModifyWeekRequest": {
        "type": "object",
        "properties": {
          "lun": {
            "$ref": "#/components/schemas/WeekDay"
          },
          "mar": {
            "$ref": "#/components/schemas/WeekDay"
          },
          "mer": {
            "$ref": "#/components/schemas/WeekDay"
          },
          "gio": {
            "$ref": "#/components/schemas/WeekDay"
          },
          "ven": {
            "$ref": "#/components/schemas/WeekDay"
          },
          "sab": {
            "$ref": "#/components/schemas/WeekDay"
          },
          "dom": {
            "$ref": "#/components/schemas/WeekDay"
          }
        }
      },
      "ItemRequest": {
        "required": [
          "description",
          "name",
          "price"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          }
        }
      },
...
      "UpdateSocialBakery": {
        "type": "object",
        "properties": {
          "facebook": {
            "type": "string"
          },
          "instagram": {
            "type": "string"
          },
          "twitter": {
            "type": "string"
          }
        }
      },
      "ReqModDesc": {
        "type": "object",
        "properties": {
          "desc": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}